requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
pydantic==2.7.1
python-dotenv==1.0.1
pytest==8.2.1
openai>=2.7.0